abstracted and consolidated logic from two yup validation forms
